requirements.txt
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
et-xmlfile==1.1.0
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
numpy==1.26.1
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.1.1
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
